Refactor import script
Break out files for the import script so it can grow to
support a new use case, running in CI on changed files,
while still supporting the old use case of batch imports.
