+++ .gitignore
@@ -22,3 +22,5 @@ import/.venv/Scripts/deactivate.bat
 import/.venv/Scripts/Activate.ps1
 import/.venv/Scripts/activate.bat
 import/.venv/Scripts/activate
+
+.idea/
