+++ import/.gitignore
@@ -3,4 +3,4 @@ bin
 lib
 lib64
 pyvenv.cfg
-
+.venv
