+++ import/config.py
@@ -1,0 +1,3 @@
+generated_dir = "generated"
+datasets_output_dir = f"{generated_dir}/_datasets"
+log_filename = f"{generated_dir}/error.log"
